+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Release 0.4 (2023-10-03)
 * Add ability to override gpiochip selection; see :ref:`gpio_chip`
 * Convert bouncetime -666 to :data:`None` (bug compatibility, which also
   ensures this should work with GPIO Zero's rpigpio pin driver)
+* Fix ``pull_up_down`` default on :func:`setup`
 
 
 Release 0.3 (2022-10-14)